skip -1 values for upstream connect and header time added
If we accessed several upstreams and some of them were unsuccessful,
the values of connect_time, header_time in such upstreams
may remain equal to the initialization value -1,
although the status of such upstreams may be non-zero.
As a result, the values of upstream connect and header time
in statistics may be less than the real values.

To fix this problem -1 values of connect and header time are skipped.